+++ src/aui-scheduler/HISTORY.md
@@ -4,7 +4,7 @@
 
 ## @VERSION@
 
-No registries yet.
+* [AUI-1772](https://issues.liferay.com/browse/AUI-1772) Improve performance of SchedulerTableView plotEvents method
 
 ## [3.0.0](https://github.com/liferay/alloy-ui/releases/tag/3.0.0)